Calculate the expected value of partial perfect information from a decision model

Calculate the expected value of partial perfect information from a decision model

evppi.remote(
  outputs,
  inputs,
  pars = NULL,
  se = FALSE,
  B = 500,
  nsim = NULL,
  verbose = FALSE,
  method = NULL,
  ...
)

Arguments

outputs
This could take one of two forms

"net benefit" form: a matrix or data frame of samples from the uncertainty distribution of the expected net benefit. The number of rows should equal the number of samples, and the number of columns should equal the number of decision options.

"cost-effectiveness analysis" form: a list with the following named components:

"c": a matrix or data frame of samples from the distribution of costs. There should be one column for each decision option.

"e": a matrix or data frame of samples from the distribution of effects, likewise.

"k": a vector of willingness-to-pay values.

Objects of class "bcea", as created by the BCEA package, are in this "cost-effectiveness analysis" format, therefore they may be supplied as the outputs argument.

If outputs is a matrix or data frame it is assumed to be of "net benefit" form. Otherwise if it is a list, it is assumed to be of "cost effectiveness analysis" form.

inputs
Matrix or data frame of samples from the uncertainty distribution of the input parameters of the decision model. The number of columns should equal the number of parameters, and the columns should be named. This should have the same number of rows as there are samples in outputs, and each row of the samples in outputs should give the model output evaluated at the corresponding parameters.

pars
A character vector giving the parameters of interest, for which a single EVPPI calculation is required. If the vector has multiple element, then the joint expected value of perfect information on all these parameters together is calculated.

Alternatively, pars may be a list. Multiple EVPPI calculations are then performed, one for each component of pars defined in the above vector form.

pars must be specified if inputs is a matrix or data frame. This should then correspond to particular columns of inputs. If inputs is a vector, this is assumed to define the single parameter of interest, then pars is not required.

se
If possible, calculate a standard error for the EVPPI. Currently only supported for method="gam".

B
Number of parameter replicates for calculating the standard error.

nsim
Number of simulations from the model to use for calculating EVPPI. The first nsim rows of the objects in inputs and outputs are used.

verbose
If TRUE, then messages are printed describing each step of the calculation, if the method supplies these. Useful to see the progress of slow calculations.

method
Character string indicating the calculation method. The default methods are based on nonparametric regression

...
Other arguments to control specific methods.

Value

A data frame with a column pars indicating the parameter(s) and a column evppi giving the corresponding EVPPI.

If outputs is of "cost-effectiveness analysis" form so that there is one EVPPI per willingness-to-pay value, then a column k identifies the willingness-to-pay.

If standard errors are requested, then the standard errors are returned in the column se.

Examples

set.seed(1)
nsam <- 10000
inputs <- data.frame(
  p1 = rnorm(nsam, 1, 1),
  p2 = rnorm(nsam, 0, 2)
)
outputs_nb <- data.frame(
  t1 = 0,
  t2 = inputs$p1 - inputs$p2
)
evppi.remote(outputs_nb, inputs, pars="p1")

#> Calling server at https://prism.peermodelsnetwork.com/route/voi/run

#> $pars
#> [1] "p1"
#> 
#> $evppi
#> [1] 0.0822
#>